pyproject.toml
[project]
name = "simple-stream-room"
version = "0.1.0"
description = "simple simultaneous synchronized transmission server for multiple users"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.1",
    "ffmpeg-python>=0.2.0",
    "python-socketio>=5.14.2",
    "uvicorn>=0.38.0",
]
